+++ src/context/example.jsx
@@ -37,6 +37,5 @@ function Provider({children}) {
     )
 }
 
-
 export {Provider};
 export default ExampleContext;
